pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slamx"
version = "0.1.0"
description = "Modern 2D LiDAR SLAM experiment platform (ROS-free core, CLI-first)"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "pyyaml>=6.0",
    "typer>=0.12",
    "scipy>=1.11",
]

[project.optional-dependencies]
rosbag = ["rosbags>=0.9"]
dev = ["pytest>=7.4", "ruff>=0.3"]

[project.scripts]
slamx = "slamx.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/slamx"]

# src layout: package directory is src/slamx, import name slamx
[tool.hatch.build.targets.sdist]
include = ["src/slamx", "tests", "configs", "examples"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
